refactor: reorganize DataList components and update exports
