+++ packages/admin-ui/src/DataList/components/MultiSelectActions.tsx
@@ -1,0 +1,11 @@
+import { DataListProps } from "../types";
+
+const MultiSelectActions = (props: DataListProps) => {
+    const { multiSelectActions } = props;
+    if (!multiSelectActions) {
+        return null;
+    }
+    return multiSelectActions;
+};
+
+export { MultiSelectActions };
