+++ packages/admin-ui/src/DataList/components/MultiSelectAll.tsx
@@ -1,0 +1,33 @@
+import React from "react";
+import { CheckboxPrimitive } from "~/Checkbox";
+import { DataListProps } from "../types";
+
+const MultiSelectAll = (props: DataListProps) => {
+    const { multiSelectActions } = props;
+    if (!multiSelectActions) {
+        return null;
+    }
+    /**
+     * We can safely cast because we have defaults.
+     */
+    const { isAllMultiSelected, isNoneMultiSelected, multiSelectAll, data } =
+        props as Required<DataListProps>;
+
+    return (
+        <>
+            {typeof multiSelectAll === "function" && (
+                <div className={"wby-size-10 wby-flex wby-items-center wby-justify-center"}>
+                    <CheckboxPrimitive
+                        indeterminate={!isAllMultiSelected(data) && !isNoneMultiSelected(data)}
+                        checked={isAllMultiSelected(data)}
+                        onCheckedChange={() => {
+                            multiSelectAll(!isAllMultiSelected(data), data);
+                        }}
+                    />
+                </div>
+            )}
+        </>
+    );
+};
+
+export { MultiSelectAll };
