--- packages/ui/src/List/DataList/DataListModalOverlay/DataListModalOverlay.tsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-export interface DataListModalOverlayProps {
-    children: React.ReactNode;
-}
-
-export const DataListModalOverlay = ({ children }: DataListModalOverlayProps) => {
-    return <div data-testid={"data-list-modal-wrapper"}>{children}</div>;
-};
