+++ packages/ui/src/List/DataList/DataListModalOverlay/index.ts
@@ -1,6 +1,15 @@
-export { DataListModalOverlay } from "./DataListModalOverlay";
-export { DataListModalOverlayAction } from "./DataListModalOverlayAction";
-export {
-    DataListModalOverlayContext,
-    DataListModalOverlayProvider
-} from "./DataListModalOverlayContext";
+import { DataListModal } from "@webiny/admin-ui";
+
+/**
+ * @deprecated This component is deprecated and will be removed in future releases.
+ * Please use the `DataListModal.Content` component from the `@webiny/admin-ui` package instead.
+ */
+const DataListModalOverlay = DataListModal.Content;
+
+/**
+ * @deprecated This component is deprecated and will be removed in future releases.
+ * Please use the `DataListModal.Trigger` component from the `@webiny/admin-ui` package instead.
+ */
+const DataListModalOverlayAction = DataListModal.Trigger;
+
+export { DataListModalOverlay, DataListModalOverlayAction };
