+++ packages/ui/src/List/DataList/index.ts
@@ -1,4 +1,3 @@
 export { DataList, ScrollList } from "./DataList";
 export { DataListWithSections, ScrollListWithSections } from "./DataListWithSections";
-export { default as NoData } from "./NoData";
 export * from "./DataListModalOverlay";
